start fixing smoke tests
