+++ tests/templates/kuttl/smoke/51-copy-metrics-test-script.yaml
@@ -2,5 +2,5 @@
 apiVersion: kuttl.dev/v1beta1
 kind: TestStep
 commands:
-  - script: kubectl cp -n $NAMESPACE ./test_metrics.py test-runner-0:/tmp
-  - script: kubectl cp -n $NAMESPACE ./test_prometheus_metrics.py test-runner-0:/tmp
+  - script: kubectl cp -n $NAMESPACE ./test_jmx_metrics.py test-runner-0:/tmp
+  - script: kubectl cp -n $NAMESPACE ./test_native_metrics.py test-runner-0:/tmp
